requirements.txt
django_taggit==1.1.0